Create dedicated test file for middleware (test_middleware.py) and add test that pins down the order that middleware is executed in
