+++ graphql/execution/tests/test_executor.py
@@ -565,111 +565,6 @@ def test_exceptions_are_reraised_if_specified(mocker):
         "An error occurred while resolving field Query.foo")
 
 
-def test_middleware():
-    doc = '''{
-        ok
-        not_ok
-    }'''
-
-    class Data(object):
-
-        def ok(self):
-            return 'ok'
-
-        def not_ok(self):
-            return 'not_ok'
-
-    doc_ast = parse(doc)
-
-    Type = GraphQLObjectType('Type', {
-        'ok': GraphQLField(GraphQLString),
-        'not_ok': GraphQLField(GraphQLString),
-    })
-
-    def reversed_middleware(next, *args, **kwargs):
-        p = next(*args, **kwargs)
-        return p.then(lambda x: x[::-1])
-
-    middlewares = MiddlewareManager(reversed_middleware)
-    result = execute(GraphQLSchema(Type), doc_ast,
-                     Data(), middleware=middlewares)
-    assert result.data == {'ok': 'ko', 'not_ok': 'ko_ton'}
-
-
-def test_middleware_class():
-    doc = '''{
-        ok
-        not_ok
-    }'''
-
-    class Data(object):
-
-        def ok(self):
-            return 'ok'
-
-        def not_ok(self):
-            return 'not_ok'
-
-    doc_ast = parse(doc)
-
-    Type = GraphQLObjectType('Type', {
-        'ok': GraphQLField(GraphQLString),
-        'not_ok': GraphQLField(GraphQLString),
-    })
-
-    class MyMiddleware(object):
-        def resolve(self, next, *args, **kwargs):
-            p = next(*args, **kwargs)
-            return p.then(lambda x: x[::-1])
-
-    middlewares = MiddlewareManager(MyMiddleware())
-    result = execute(GraphQLSchema(Type), doc_ast,
-                     Data(), middleware=middlewares)
-    assert result.data == {'ok': 'ko', 'not_ok': 'ko_ton'}
-
-
-def test_middleware_skip_promise_wrap():
-    doc = '''{
-        ok
-        not_ok
-    }'''
-
-    class Data(object):
-
-        def ok(self):
-            return 'ok'
-
-        def not_ok(self):
-            return 'not_ok'
-
-    doc_ast = parse(doc)
-
-    Type = GraphQLObjectType('Type', {
-        'ok': GraphQLField(GraphQLString),
-        'not_ok': GraphQLField(GraphQLString),
-    })
-
-    class MyPromiseMiddleware(object):
-        def resolve(self, next, *args, **kwargs):
-            return Promise.resolve(next(*args, **kwargs))
-
-    class MyEmptyMiddleware(object):
-        def resolve(self, next, *args, **kwargs):
-            return next(*args, **kwargs)
-
-    middlewares_with_promise = MiddlewareManager(
-        MyPromiseMiddleware(), wrap_in_promise=False)
-    middlewares_without_promise = MiddlewareManager(
-        MyEmptyMiddleware(), wrap_in_promise=False)
-
-    result1 = execute(GraphQLSchema(Type), doc_ast, Data(),
-                      middleware=middlewares_with_promise)
-    result2 = execute(GraphQLSchema(Type), doc_ast, Data(),
-                      middleware=middlewares_without_promise)
-    assert result1.data == result2.data and result1.data == {
-        'ok': 'ok', 'not_ok': 'not_ok'}
-
-
 def test_executor_properly_propogates_path_data(mocker):
     time_mock = mocker.patch('time.time')
     time_mock.side_effect = range(0, 10000)
